Split hpx_wrap from hpx_init again
