+++ cmake/HPX_GeneratePackage.cmake
@@ -45,7 +45,7 @@ install(EXPORT HPXTargets
 
 if (NOT MSVC)
   add_library(hpx_pkgconfig_application INTERFACE)
-  target_link_libraries(hpx_pkgconfig_application INTERFACE hpx_init)
+  target_link_libraries(hpx_pkgconfig_application INTERFACE hpx hpx_init hpx_wrap)
   target_compile_definitions(hpx_pkgconfig_application INTERFACE HPX_APPLICATION_EXPORTS)
   target_compile_options(hpx_pkgconfig_application INTERFACE "-std=c++${HPX_CXX_STANDARD}")
